pyproject.toml
[project]
name = "vmboot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ansible>=11.3.0",
    "jinja2>=3.1.6",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]
